refactor: sugaming-site layout structure
Refactored the application layout structure, so that less code is repeated.
